refactor dialogs to separate file
